pyproject.toml
[project]
name = "crewaix"
version = "0.1.0"
description = "CrewAI Tweet Generator"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "crewai>=1.8.0",
    "litellm==1.75.0",
    "python-dotenv>=1.1.1",
    "pydantic[email]>=2.11.0",
    "pyyaml>=6.0.0",
]
